Refactor to support multiple contexts to separate files
