+++ src/Exceptions/CredentialValidationException.php
@@ -13,3 +13,4 @@ namespace GettyImages\Api\Exception {
     {
     }
 }
+
